# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chore-Ganizer API",
    "description": "Family-friendly chore management system API",
    "version": "1.0.0",
    "contact": {
      "name": "Chore-Ganizer"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative path (via nginx proxy)"
    },
    {
      "url": "http://localhost:3010/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "description": "Retrieve all users (Parents only)",
        "operationId": "getAllUsers",
        "security": [
          { "cookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized (Parent role required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID (Parents only)",
        "operationId": "getUserById",
        "security": [
          { "cookieAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized (Parent role required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/chore-assignments/upcoming": {
      "get": {
        "tags": ["Chore Assignments"],
        "summary": "Get upcoming assignments",
        "description": "Get assignments due in the next N days",
        "operationId": "getUpcomingAssignments",
        "security": [
          { "cookieAuth": [] }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look ahead (default: 7)",
            "schema": {
              "type": "integer",
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of upcoming assignments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChoreAssignmentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chore-assignments/overdue": {
      "get": {
        "tags": ["Chore Assignments"],
        "summary": "Get overdue assignments",
        "description": "Get all overdue (past due date) assignments",
        "operationId": "getOverdueAssignments",
        "security": [
          { "cookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "List of overdue assignments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChoreAssignmentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/chore-categories/{id}/templates": {
      "get": {
        "tags": ["Chore Categories"],
        "summary": "Get templates by category",
        "description": "Retrieve all templates in a specific category",
        "operationId": "getCategoryTemplates",
        "security": [
          { "cookieAuth": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Chore Category ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of templates in the category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChoreTemplateListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "dad@home"
          },
          "name": {
            "type": "string",
            "example": "Dad"
          },
          "role": {
            "type": "string",
            "enum": ["PARENT", "CHILD"],
            "example": "PARENT"
          },
          "points": {
            "type": "integer",
            "example": 150
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "users": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Error message"
              },
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              }
            }
          }
        }
      },
      "ChoreTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Wash dishes"
          },
          "description": {
            "type": "string",
            "example": "Wash all dishes after dinner"
          },
          "points": {
            "type": "integer",
            "example": 10
          },
          "icon": {
            "type": "string",
            "example": "soap"
          },
          "color": {
            "type": "string",
            "example": "#FF6B6B"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChoreAssignment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "templateId": {
            "type": "integer",
            "example": 1
          },
          "template": {
            "$ref": "#/components/schemas/ChoreTemplate"
          },
          "assignedToId": {
            "type": "integer",
            "example": 2
          },
          "assignedTo": {
            "$ref": "#/components/schemas/User"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-31T23:59:59.000Z"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "COMPLETED"],
            "example": "PENDING"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isOverdue": {
            "type": "boolean",
            "description": "Computed field: true if past due date and not completed",
            "example": false
          }
        }
      },
...
      "ChoreTemplateListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "templates": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ChoreTemplate"
                }
              }
            }
          }
        }
      },
...
      "ChoreAssignmentListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "assignments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ChoreAssignment"
                }
              }
            }
          }
        }
      },
...
    }
  }
}